patch carte

+++ src/main/java/fr/miage/bank/input/CarteInput.java
@@ -29,7 +29,7 @@ public class CarteInput {
 
     @NotNull
     @Min(0)
-    private int plafond;
+    private double plafond;
 
     @NotNull
     private boolean sansContact;
